Don't consider an order risky until we know it is actually risky.
The risk analysis box shows up in admin before we have even heard called
out to or heard back from signifyd. During my appointment with a guide yesterday,
he said this is a frequent point of confusion for customers that they
are checking out, and causes the appointment to end on a sour note.