Add option to skip store credit orders
In the event that an order is completely paid by store credit, it makes
little sense to send to SIGNIFYD if you have to pay them for every order
they process. For those users we offer the option to immediately approve
orders which have only store credit payments and avoid creating a
SIGNIFYD case entirely.